# front-end/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LittyPicky API",
    "description": "A gamified litter-picking community app with geospatial features",
    "contact": {
      "name": "LittyPicky Team",
      "email": "team@littypicky.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/reports": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all reports (not just nearby)",
        "description": "GET /api/admin/reports",
        "operationId": "list_all_reports",
        "responses": {
          "200": {
            "description": "Returns all reports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminReportView"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/reports/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a report (for spam/inappropriate content)",
        "description": "DELETE /api/admin/reports/:id",
        "operationId": "delete_report",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Report ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Report not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all users (paginated)",
        "description": "GET /api/admin/users?page=1&limit=20",
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "Returns list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get user by ID",
        "description": "GET /api/admin/users/:id",
        "operationId": "get_user_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/users/{id}/ban": {
      "put": {
        "tags": [
          "Admin"
        ],
        "operationId": "toggle_user_ban",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User ban status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminReportView": {
        "type": "object",
        "required": [
          "id",
          "reporter_id",
          "latitude",
          "longitude",
          "status",
          "created_at",
          "reporter_name",
          "reporter_email"
        ],
        "properties": {
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "claimed_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "cleared_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cleared_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "reporter_email": {
            "type": "string"
          },
          "reporter_id": {
            "type": "string",
            "format": "uuid"
          },
          "reporter_name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReportStatus"
          }
        }
      },
...
      "BanUserRequest": {
        "type": "object",
        "description": "Ban/unban a user\nPUT /api/admin/users/:id/ban",
        "required": [
          "is_active"
        ],
        "properties": {
          "is_active": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "ReportStatus": {
        "type": "string",
        "enum": [
          "pending",
          "claimed",
          "cleared",
          "verified"
        ]
      },
...
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "city",
          "country",
          "search_radius_km",
          "role",
          "email_verified",
          "created_at"
        ],
        "properties": {
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "full_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "search_radius_km": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}